lol xd bug fixers

+++ functions/autoVC/vcchecker.js
@@ -1,7 +1,6 @@
 module.exports.exec = (v, lol, prVC, link, sv) => {
     let temp = lol;
     let vcchan = temp.vc;
-    if(vcchan.id == sv.mainvc) return 42;
     const [members, bots] = vcchan.members.partition(m => m.user.bot == false);
     if(members.size == 0) {
         if(link.has(temp.vc.id)) {
@@ -9,6 +8,7 @@ module.exports.exec = (v, lol, prVC, link, sv) => {
             text.delete();
             link.delete(temp.vc.id);
         }
+        if(vcchan.id == sv.mainvc) return [false, temp.vc.id]
         prVC.delete(v.member.id);
         vcchan.delete();
         return [true, temp.vc.id];
